[wanglei] ref: 没有好的思路，只修改了一处

+++ 3/src/main/java/com/twu/refactoring/Direction.java
@@ -51,7 +51,7 @@ public class Direction {
 
     @Override
     public int hashCode() {
-        return (int) direction;
+        return direction;
     }
 
     @Override
